Don't download deps if they are already in buildfolder/bin/*

+++ backend/eks/deps.sh
@@ -9,11 +9,16 @@ if [[ "$DOWNLOAD_BINS" == "false" ]]; then
 fi
 
 
-# pin the kubectl to eks default version. Hardcoded as the URL has a changing date stamp
-curl -o kubectl https://amazon-eks.s3-us-west-2.amazonaws.com/1.14.6/2019-08-22/bin/linux/amd64/kubectl
-chmod +x kubectl && mv kubectl bin/
 
-if ! which aws ; then
+awskubectlpath=bin/kubectl
+if [ ! -e "$awskubectlpath" ]; then
+    # pin the kubectl to eks default version. Hardcoded as the URL has a changing date stamp
+    curl -o kubectl https://amazon-eks.s3-us-west-2.amazonaws.com/1.14.6/2019-08-22/bin/linux/amd64/kubectl
+    chmod +x kubectl && mv kubectl bin/
+fi
+
+awspath=bin/aws
+if [ ! -e "$awspath" ]; then
     mkdir -p .local
     curl "https://s3.amazonaws.com/aws-cli/awscli-bundle.zip" -o "awscli-bundle.zip"
     unzip awscli-bundle.zip
@@ -24,8 +29,10 @@ if ! which aws ; then
     chmod +x aws-iam-authenticator && mv aws-iam-authenticator bin/
 fi
 
-curl -o terraform.zip https://releases.hashicorp.com/terraform/0.12.9/terraform_0.12.9_linux_amd64.zip
-# curl -o terraform.zip https://releases.hashicorp.com/terraform/0.11.14/terraform_0.11.14_linux_amd64.zip
-unzip terraform.zip
-chmod +x terraform && mv terraform bin/
-rm -rf terraform.zip
+terraformpath=bin/terraform
+if [ ! -e "$terraformpath" ]; then
+    curl -o terraform.zip https://releases.hashicorp.com/terraform/0.12.9/terraform_0.12.9_linux_amd64.zip
+    # curl -o terraform.zip https://releases.hashicorp.com/terraform/0.11.14/terraform_0.11.14_linux_amd64.zip
+    unzip terraform.zip && rm -rf terraform.zip
+    chmod +x terraform && mv terraform bin/
+fi
