+++ backend/minikube/deps.sh
@@ -1,22 +1,26 @@
 #!/bin/bash
 
-# duplicated in s/include/common.sh, needed for bootstrapping:
-export CLUSTER_NAME=${CLUSTER_NAME:-minikube}
-
 . ../../include/common.sh
+. .envrc
 
 
 MINIKUBE_VERSION=latest
-
 if [[ "$OSTYPE" == "darwin"* ]]; then
     MINIKUBE_BIN=minikube-darwin-amd64
 else
     MINIKUBE_BIN=minikube-linux-amd64
 fi
 
-curl -Lo minikube https://storage.googleapis.com/minikube/releases/"$MINIKUBE_VERSION"/"$MINIKUBE_BIN"
-chmod +x minikube && mv minikube bin/
-curl -Lo docker-machine-driver-kvm2 https://storage.googleapis.com/minikube/releases/latest/docker-machine-driver-kvm2
-chmod +x docker-machine-driver-kvm2 && mv docker-machine-driver-kvm2 bin/
+minikubepath=bin/minikube
+if [ ! -e "$minikubepath" ]; then
+    curl -Lo minikube https://storage.googleapis.com/minikube/releases/"$MINIKUBE_VERSION"/"$MINIKUBE_BIN"
+    chmod +x minikube && mv minikube bin/
+fi
+
+dockermachinedriverkvm2path=bin/docker-machine-driver-kvm2
+if [ ! -e "$dockermachinedriverkvm2path" ]; then
+    curl -Lo docker-machine-driver-kvm2 https://storage.googleapis.com/minikube/releases/latest/docker-machine-driver-kvm2
+    chmod +x docker-machine-driver-kvm2 && mv docker-machine-driver-kvm2 bin/
+fi
 
 popd || exit
